requirements.txt
# 桥梁裂纹智能检测系统依赖
# Bridge Crack Identification System Dependencies

# 核心科学计算
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.4.0

# 数据处理
pyyaml>=5.4.0

# 可选：用于加速计算的库
# numba>=0.54.0
torch>=2.0.0
